Use 'find_pkg_directory' for pack
